// src/lib/viz/generated/map-rooms.generated.ts
export const roomDataUnscaled = {
    rooms: [
        {
            sceneName: 'Crossroads_01',
            spriteInfo: {
                name: 'Crossroads_01',
                size: {
                    x: 156.0,
                    y: 73.0,
                },
                padding: {
                    x: 0.0,
                    y: 0.0,
                    z: 0.0,
                    w: 0.0,
                },
            },
// ...
        },
// ...
};